src/backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
cachetools==5.4.0
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
filelock==3.16.0
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.9.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
groq==0.11.0
gspread==6.1.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
interchange==2021.0.4
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-core==0.3.1
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.121
llvmlite==0.43.0
Mako==1.3.5
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python-headless
orjson==3.10.7
packaging==24.1
paho-mqtt==2.1.0
pansi==2020.7.3
pillow==10.4.0
ping3==4.0.8
proto-plus==1.24.0
protobuf==5.27.2
psutil==6.0.0
py2neo==2021.2.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.2
pyTelegramBotAPI==4.19.2
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.1.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.4
SQLAlchemy==2.0.31
srt==3.5.3
sympy==1.13.2
telebot==0.0.5
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
vosk==0.3.45
websockets==13.0.1
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
yarl==1.11.1
